js: sorted imports & moved constants to js/settings.js
